Int to float CreatePrimary ray
change int values X and Y to create a primary ray to float values X and
Y

+++ raytracergraphics-master/raytracer/Camera.cs
@@ -68,7 +68,7 @@ namespace Template
         {
             get { return (position + appliedFieldOfView * viewDirection); }
         }
-        public Ray CreatePrimaryRay(int x, int y)
+        public Ray CreatePrimaryRay(float x, float y)
         {
             float u = x * displayWidth*2;
             float v = y * displayHeight;
